fix for casting of bf16 type

+++ include/matx/operators/scalar_ops.h
@@ -343,8 +343,7 @@ template <typename T1, typename T2> struct SubF {
   {
     if constexpr (is_complex_v<T1> && std::is_arithmetic_v<T2>) {
       if constexpr (is_complex_half_v<T1>) {
-        return (T1){v1.real() - static_cast<typename T1::value_type>(
-                                    static_cast<float>(v2)),
+        return (T1){v1.real() - static_cast<typename T1::value_type>(static_cast<float>(v2)),
                     v1.imag() };
       }
       else {
@@ -354,11 +353,11 @@ template <typename T1, typename T2> struct SubF {
     }
     else if constexpr (is_complex_v<T2> && std::is_arithmetic_v<T1>) {
       if constexpr (is_complex_half_v<T2>) {
-        return (T2){static_cast<typename T2::value_type>(static_cast<float>(v1) - v2.real() ),
+        return (T2){static_cast<typename T2::value_type>(static_cast<float>(v1) - static_cast<float>(v2.real()) ),
                     -v2.imag() };
       }
       else {
-        return (T2){static_cast<typename T2::value_type>(v1) - v2.real() ,
+        return (T2){static_cast<typename T2::value_type>(v1) - v2.real(),
                     -v2.imag() };
       }
     }
